feat: :construction: Each instance of background is now a different one.
Now you can have different backgrounds on your screens, it should work fine with two screens but not so sure about three or more. It sould have a bg but not a different one that the second screen one. Will do some cleaning and improvements tomorrow.

External screns are now different instances.